Closes #2 - Non-owner of the image can delete the image

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -147,9 +147,18 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session, Model model) {
+        Image image = imageService.getImage(imageId);
+        User user = (User) session.getAttribute("loggeduser");
+        if (image.getUser() != null && image.getUser().getId().equals(user.getId())) {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
+        String error = "Only the owner of the image can delete the image";
+        model.addAttribute("deleteError", error);
+        model.addAttribute("image", image);
+        model.addAttribute("tags", image.getTags());
+        return "images/image";
     }
